[move] GraphNetworkを移動
